切割比例导入，删除修改bug

+++ src/main/java/com/loushuiyifan/report/bean/RptImportDataCut.java
@@ -21,5 +21,5 @@ public class RptImportDataCut {
     String activeFlag;
     String chgWho;
     Date lstUpd;
-    Integer groupId;
+    Long groupId;
 }
